Add images_hierarchy.py file
