+++ tagging/apply_tags.py
@@ -3,61 +3,11 @@
 # Distributed under the terms of the Modified BSD License.
 import argparse
 import logging
-from dataclasses import dataclass, field
-from typing import Optional, List
-from taggers import TaggerInterface, \
-    SHATagger, \
-    PythonVersionTagger, \
-    JupyterNotebookVersionTagger, JupyterLabVersionTagger, JupyterHubVersionTagger, \
-    RVersionTagger, TensorflowVersionTagger, JuliaVersionTagger, \
-    SparkVersionTagger, HadoopVersionTagger, JavaVersionTagger
 from plumbum.cmd import docker
+from images_hierarchy import ALL_IMAGES
 
 
 logger = logging.getLogger(__name__)
-
-
-@dataclass
-class ImageDescription:
-    parent_image: Optional[str]
-    taggers: List[TaggerInterface] = field(default_factory=list)
-
-
-ALL_IMAGES = {
-    "base-notebook": ImageDescription(
-        parent_image=None,
-        taggers=[
-            SHATagger,
-            PythonVersionTagger, JupyterNotebookVersionTagger, JupyterLabVersionTagger, JupyterHubVersionTagger
-        ]
-    ),
-    "minimal-notebook": ImageDescription(
-        parent_image="base-notebook"
-    ),
-    "scipy-notebook": ImageDescription(
-        parent_image="minimal-notebook"
-    ),
-    "r-notebook": ImageDescription(
-        parent_image="minimal-notebook",
-        taggers=[RVersionTagger]
-    ),
-    "tensorflow-notebook": ImageDescription(
-        parent_image="scipy-notebook",
-        taggers=[TensorflowVersionTagger]
-    ),
-    "datascience-notebook": ImageDescription(
-        parent_image="scipy-notebook",
-        taggers=[JuliaVersionTagger]
-    ),
-    "pyspark-notebook": ImageDescription(
-        parent_image="scipy-notebook",
-        taggers=[SparkVersionTagger, HadoopVersionTagger, JavaVersionTagger]
-    ),
-    "allspark-notebook": ImageDescription(
-        parent_image="pyspark-notebook",
-        taggers=[RVersionTagger]
-    )
-}
 
 
 def get_all_taggers(short_image_name):
